map open blocks to LaTeX environments
The converter now maps open blocks to LaTeX environments.
See the README, section 4.3, "LaTeX Environments" for details.
Briefly, a file, "new_environments.tex" is generated which
contains a simple but serviceable definition of a LaTeX
environment corresponding to the given open block.  These
definitions are stored in the file "new_environents.tex"
The user can use this file as a starting point for customized
environments.