Modifying environment_block and converter to use caption for automatic numbering of theorems, equations, etc. per the suggestion of @mojavelinux.  (Works like a charm, 25+ lines of bad code eliminated).
HOWEVEr, I am stuck on lines 121-125 of converter.rb.  In line 123 I edit
the document.references map.  The changes made are confirmed
in the logging statement of line 124.  However, when the next
equation, theorem, etc is encountered, the new information
added is overwritten with default information.  I must be
executing the code of line #123 at the wrong moment.  However,
I don't see where else I can excecute #123.

Run

  $ asciidoctor -r asciidoctor-latex -a stem=latexmath -b html try-out/eqno-latex.adoc

in the root of this repo to see the problem.