Added module TeXUtlities to faciltate constructon of LaTeX strings from Ascidoctor nodes.  For example, the string
  "\\begin\{aquote\}{#{attribution}#{citetitle ? ' - ' + citetitle : ''}}\n#{self.content}\\end\{aquote\}\n"

is replaced by

  $tex.env 'aquote', attribution, citetitle, self.content

Things are set up within module Asciidoctor, file
`nodeprocessors.rb` as

  include TexUtilities
  $tex = TexUtilities

@jirutka or @mojavelinux could you advise me on this?
I'd like to get rid of that pesky global variable but
am unsure of how to proceed. Is this approach OK with you?
(See also core_ext/utility.rb for some related Monkeypatching).

The other methods of note in the TexUtilities package are:

  $tex.macro 'foo-sez', 'ho ho ho!'
  => "\\foo-sez{ho ho ho!}"

  tex.apply_macros ['foo-sez', 'shout!'], 'ho ho ho!'
  => "\\foo-sez{\\shout!{ho ho ho!}}"