Do not enclose the preamble in \begin{preamble} ... \end{preamble}
The preamble consists of all content between the document header
and the first section.  In the source text this can be empty,
or it can consist of several paragraphs of text.  However,
this is also the region in which the  [env.texmacro] block
occurs.  It cannot be placed inside formatting elements.
For now we leave the preamble 'unprocessed' in the LaTeX
backend, except to insert the content of the env.texmacro
block.