Quick fix for manual TOC placement
You can now position the toc manually using the `toc` macro as
documented by the User Manual.
However it is not a full implement of the `toc` macro because it is very
hard to do in LaTeX.
Indeed, the formatting of `\tableofcontents` is decided in the preamble
whereas Asciidoctor defines it when calling the macro.